+++ extensions/omni_httpd/CHANGELOG.md
@@ -11,6 +11,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * HTTP response should contain correct status code reasons [#781](https://github.com/omnigres/omnigres/pull/781)
 
+### Changed
+
+* Core routing changed to a router mechanism with URLPattern support [#783](https://github.com/omnigres/omnigres/pull/783)
+
 ## [0.3.0] - 2025-02-04
 
 ### Added